\name{blp}
\title{Deprecated.}
\usage{blp(conn, securities, fields, start, end, barsize,
    barfields, retval, override_fields, overrides,
    currency)}

\description{
  Deprecated.
}

\details{
  \code{blp} and \code{blpGetData} are deprecated in favour
  of the functions \code{\link{bdp}}, \code{\link{bdh}},
  \code{\link{tick}}, \code{\link{bar}}.
}
\alias{blp}
\alias{blpGetData}
