\name{blpDisconnect}
\alias{blpDisconnect}
\title{Close the bloomberg connection.}
\usage{blpDisconnect(conn)}

\description{
  Close the bloomberg connection.
}

\details{
  Closes the connection and performs a round of garbage
  collection.
}
\seealso{\code{\link{blpConnect}}}
\arguments{
  \item{conn}{a Bloomberg connection object}
}
