\name{blpFieldInfo}
\alias{blpFieldInfo}
\title{Get the field metadata from the bloomberg data dictionary...}
\usage{blpFieldInfo(conn, fields)}

\description{
  Get the field metadata from the bloomberg data dictionary
}
\seealso{\code{\link{field.description}}}
\arguments{
  \item{conn}{a Bloomberg connection object}
  \item{fields}{a vector of field mnemonics}
}
