\name{field.description}
\alias{field.description}
\title{Get the field descriptions from the bloomberg data dictionary...}
\usage{field.description(conn, fields)}

\description{
  Get the field descriptions from the bloomberg data
  dictionary
}
\seealso{\code{\link{blpFieldInfo}}}
\arguments{
  \item{conn}{a Bloomberg connection object}
  \item{fields}{a vector of field mnemonics}
}
