\name{RBloomberg}
\title{R interface to the Bloomberg API.}

\description{
  R interface to the Bloomberg API.
}

\details{
  RBloomberg provides R wrappers for the Bloomberg v3 API.
  The functions \code{\link{bdp}}, \code{\link{bds}},
  \code{\link{bdh}}, \code{\link{bar}}, and
  \code{\link{tick}} are implemented.  All return results
  as data frames.

  Connections are created with \code{\link{blpConnect}} and
  use the Java API.
}
\docType{package}
\alias{RBloomberg}
\alias{package-RBloomberg}
