\name{blpConnect}
\alias{blpConnect}
\alias{blpDisconnect}
\title{Open or shut a connection to Bloomberg}
\description{
  The function \code{blpConnect} returns a connection object to one of
  the Bloomberg interfaces. Currently, only the Java Version 3 API is
  supported.
}
\usage{
blpConnect(iface = "Java", log.level = "warning", blpapi.jar.file = NULL, throw.ticker.errors = TRUE, jvm.params = NULL, verbose = TRUE)
blpConnect.Java(log.level, blpapi.jar.file, throw.ticker.errors, jvm.params, verbose)
blpDisconnect(conn)
}
\arguments{
  \item{iface}{character. Which Bloomberg interface? (only "Java" currently supported).}
  \item{log.level}{character. Level of logs to be generated by Java. Acceptable values: warning, info, fine, finest.}
  \item{blpapi.jar.file}{character. Location of the Bloomberg API JAR file. If NULL (default), the function looks in usual locations.}
  \item{throw.ticker.errors}{logical. Should errors be thrown for invalid tickers?}
  \item{jvm.params = NULL}{Parameters passed to the Java Virtual Machine}
  \item{verbose = TRUE}{logical. Should output be written to the console on connection?}
  \item{conn}{Connection object}
}
\details{
  Connections to other APIs may be developed in future.
}
\keyword{math}
\examples{
\dontrun{conn <- blpConnect()
blpDisconnect(conn)}
}
\author{Ana Nelson \email{ana@ananelson.com}}

